Split `client-request` into several JS files
